Backend/requirements.txt
fastapi
python-dotenv
pymongo
pydantic
gunicorn
uvicorn
